requirements.txt
django
environs
djangorestframework
django-allauth
django-jazzmin
drf-spectacular
django-cors-headers
